+++ packages/design-system/hooks/native/index.ts
@@ -4,3 +4,6 @@ export { default as useColorMode } from './useColorMode'
 
 export { default as useLogoOverrides } from './useLogoOverrides'
 export type { LogoOverrides, LogoOverridesKeys } from './useLogoOverrides/types'
+
+export { default as useViewPhotoLibrary } from './useViewPhotoLibrary'
+export type { ViewPhotoLibraryProps } from './useViewPhotoLibrary/types'